feat #101: 채팅방 정보 - Mysql, 채팅 메시지 - MongoDB
