+++ Chatting-Server/src/main/java/swm/hkcc/chat/app/modules/chat/domain/ChatRoom.java
@@ -1,0 +1,28 @@
+package swm.hkcc.chat.app.modules.chat.domain;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.*;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+public class ChatRoom {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long roomId;
+
+    private String roomName;
+
+    private String roomType; // 임시
+
+    private Long missionId;
+
+    private String processStatus; // 임시
+}
+
